Style Logout Header
