+++ src/components/Navbar.js
@@ -3,11 +3,11 @@ import { Link } from 'react-router-dom';
 
 const Navbar = () => (
   <>
-    <div>
-      <Link to="/dashboard">Dashboard</Link>
-      <Link to="/tracks">Track It</Link>
-      <Link to="/progress">Progress</Link>
-      <Link to="/more">More</Link>
+    <div className="navbar">
+      <Link className="nav-item" to="/dashboard">Dashboard</Link>
+      <Link className="nav-item" to="/tracks">Track It</Link>
+      <Link className="nav-item" to="/progress">Progress</Link>
+      <Link className="nav-item" to="/more">More</Link>
     </div>
   </>
 );
