+++ src/containers/Logout.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useHistory } from 'react-router-dom';
 import { fetchLogout } from '../redux/thunk/thunkDelete';
+import '../styles/containers/Logout.scss';
 
 const Logout = () => {
   const history = useHistory();
@@ -11,10 +12,10 @@ const Logout = () => {
   };
 
   return (
-    <div>
+    <header className="logout-header">
       <h1>Tri Tracker</h1>
-      <button type="button" onClick={handleLogout}>Logout</button>
-    </div>
+      <button className="logout-btn" type="button" onClick={handleLogout}>Logout</button>
+    </header>
   );
 };
