+++ src/routes/Routes.js
@@ -4,8 +4,8 @@ import {
 } from 'react-router-dom';
 import Dashboard from '../components/Dashboard';
 import Home from '../components/Home';
-import Login from '../components/Login';
-import Register from '../components/Register';
+import Login from '../containers/Login';
+import Register from '../containers/Register';
 import Tracks from '../components/Tracks';
 
 const Routes = () => (
