+++ src/styles/components/Home.scss
@@ -2,7 +2,7 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  padding: 3rem;
+  padding: 2rem;
   font-family: 'Helvetica Neue', sans-serif;
   font-weight: bold;
   background-color: var(--primary-color);
